data-service/requirements.txt
pandas==1.3.1
scikit-learn==0.24.2
psycopg2-binary==2.9.1
python-dotenv==0.19.0